src/requirements.txt
requests==2.25.1
discord.py==1.7.3
discord-py-slash-command==1.2.2